optimization: deferred rendering
So instead of asking the event thread to render on a pixel map, let's store al render commands in a queue for the render thread to process.

This sounds like a good idea until you realize the current pixel map setup is a more efficient queue than a linked list.

Since the queue is dynamically allocated, I suspect that the cost of appending to the queue is equivalent to the cost of drawing on the staging pixel map, thus there are no gains. This just creates work for the render thread.

I leave this branch as a piece of cyf-ui's history and just in case future profiling proves me wrong.