added some test components and some initial tests
Added TEST_PUBLIC macro to indicate that a certain `private` or `protected` shall be ignored when compiled for testing purposes (ie: with ENABLE_TESTING defined)

This integrates with the CTest test suite such that it can poke and probe everything. In this case the file `test/common/common.h` defines the ENABLE_TESTING macro before including `cydui.hpp`.